requirement.txt
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
asttokens==2.0.5
async-generator==1.10
attrs==21.4.0
Babel==2.6.0
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==5.0.0
blis==0.7.7
cached-property==1.5.2
cachetools==5.0.0
catalogue==2.0.7
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
chromedriver==2.24.1
click==8.1.3
creds==1.0.15
cryptography==36.0.2
cymem==2.0.6
debugpy==1.6.0
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.6
Django==4.1.7
docutils==0.16
ebaysdk==2.1.5
entrypoints==0.4
exceptiongroup==1.1.0
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.8.0
freezegun==0.3.15
fuzzywuzzy==0.18.0
geographiclib==2.0
geopy==2.3.0
gevent==20.9.0
google-auth==2.6.6
google-auth-oauthlib==0.5.1
greenlet==0.4.17
gspread==5.3.2
h11==0.14.0
idna==2.8
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
isodate==0.6.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.5.1
jupyter==1.0.0
jupyter-client==7.3.1
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
langcodes==3.3.0
libsass==0.18.0
lxml==4.6.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
murmurhash==1.0.7
mysql-connector==2.2.9
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
nltk==3.7
notebook==6.4.11
num2words==0.5.6
numpy==1.22.3
oauthlib==3.2.0
ofxparse==0.19
outcome==1.2.0
packaging==23.0
panda==0.3.1
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.2
pathy==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
platformdirs==2.5.2
polib==1.1.0
preshed==3.0.6
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==4.21.6
psutil==5.6.6
psycopg2==2.9.3
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.8.2
pydot==1.4.1
Pygments==2.12.0
pyOpenSSL==19.0.0
pyparsing==3.0.7
PyPDF2==1.26.0
pyrsistent==0.18.1
pyserial==3.4
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-ldap==3.2.0
python-stdnum==1.13
pytz==2022.1
pyusb==1.0.2
pyzmq==22.3.0
qrcode==6.1
qtconsole==5.3.0
QtPy==2.1.0
regex==2022.4.24
reportlab==3.5.59
requests==2.22.0
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rsa==4.8
selenium==4.8.2
Send2Trash==1.8.0
six==1.16.0
smart-open==5.2.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.1
spacy==3.3.0
spacy-legacy==3.0.9
spacy-loggers==1.0.2
sqlparse==0.4.3
srsly==2.4.3
stack-data==0.2.0
terminado==0.13.3
thinc==8.0.15
tinycss2==1.1.1
tornado==6.1
tqdm==4.65.0
traitlets==5.1.1
trio==0.22.0
trio-websocket==0.9.2
typer==0.4.1
typing_extensions==4.2.0
urllib3==1.25.11
virtualenv==20.16.5
vobject==0.9.6.1
wasabi==0.9.1
wcwidth==0.2.5
webdriver-manager==3.8.5
webencodings==0.5.1
Werkzeug==0.16.1
widgetsnbextension==3.6.0
wsproto==1.2.0
xlrd==1.2.0
XlsxWriter==1.1.2
xlwt==1.3.0
zeep==3.4.0
zope.event==4.5.0
zope.interface==5.4.0
